Improve tasks

+++ tasks.py
@@ -1,6 +1,8 @@
 import os
-
+import platform
 from invoke import task
+from functools import wraps
+
 
 # Parameterizing tasks with invoke
 # https://docs.pyinvoke.org/en/0.11.1/getting_started.html#parameterizing-tasks
@@ -17,7 +19,10 @@ def t(c):
     """
     Train Rasa bot
     """
-    c.run(f"rasa train", pty=True)
+    if platform.system() != "Windows":
+        c.run(f"rasa train", pty=True)
+    else:
+        c.run(f"rasa train")
 
 
 @task
@@ -25,7 +30,10 @@ def s(c):
     """
     Rasa shell
     """
-    c.run(f"rasa shell", pty=True)
+    if platform.system() != "Windows":
+        c.run(f"rasa shell", pty=True)
+    else:
+        c.run(f"rasa shell")
 
 
 @task
@@ -33,13 +41,20 @@ def sh(c):
     """
     Rasa shell and server
     """
-    c.run(f"rasa run actions & rasa shell", pty=True)
+    if platform.system() != "Windows":
+        c.run(f"rasa run actions & rasa shell", pty=True)
+    else:
+        c.run(f"rasa run actions & rasa shell")
+
 
 
 @task
 def sv(c):
     """ Start Rasa server """
-    c.run(f"rasa run actions", pty=True)
+    if platform.system() != "Windows":
+        c.run(f"rasa run actions", pty=True)
+    else:
+        c.run(f"rasa run actions")
 
 
 @task
@@ -47,8 +62,13 @@ def stop(c):
     """
     Stop Rasa server
     """
-    c.run("pkill -f rasa", pty=True)
-    print("Rasa server stopped.")
+    if platform.system() != "Windows":
+        c.run("pkill -f rasa", pty=True)
+        print("Rasa server stopped.")
+    else:
+        # hope it works ok for windows :(
+        c.run("tskill -f rasa")
+        print("Rasa server stopped.")
 
 
 @task
@@ -56,9 +76,12 @@ def dm(c):
     """
     Remove all models on models folder
     """
-    c.run("rm -f models/*", pty=True)
-    print("All model files removed.")
-
+    if platform.system() != "Windows":
+        c.run("rm -f models/*", pty=True)
+        print("All model files removed.")
+    else:
+        c.run("rd /s /q models/*")
+        print("All model files removed.")
 
 @task
 def dt(c):
